[#312] organize folder structure
