+++ objects/jobs/job_recorded.rb
