+++ objects/log.rb
@@ -21,9 +21,8 @@
 require 'base64'
 require 'nokogiri'
 require 'aws-sdk-dynamodb'
-
-require_relative '../version'
 require_relative 'dynamo'
+require_relative '../version'
 
 #
 # Log.
