+++ objects/storage/cached_storage.rb
