+++ objects/storage/logged_storage.rb
