+++ objects/storage/once_storage.rb
