+++ objects/storage/versioned_storage.rb
