+++ objects/tickets/emailed_tickets.rb
