+++ objects/tickets/sentry_tickets.rb
@@ -20,8 +20,8 @@
 
 require 'mail'
 require 'raven'
-require_relative 'user_error'
-require_relative 'truncated'
+require_relative '../user_error'
+require_relative '../truncated'
 
 #
 # Tickets that report to Sentry.
