+++ objects/tickets/tagged_tickets.rb
