+++ objects/vcs/github.rb
@@ -19,7 +19,6 @@
 # SOFTWARE.
 
 require 'octokit'
-
 require_relative 'base'
 require_relative '../git_repo'
 require_relative '../clients/gitlab'
