+++ test/test_github_tickets.rb
@@ -22,7 +22,7 @@ require 'test/unit'
 require 'nokogiri'
 require 'yaml'
 require_relative 'test__helper'
-require_relative '../objects/tickets'
+require_relative '../objects/tickets/tickets'
 
 # GithubTickets test.
 # Author:: Yegor Bugayenko (yegor256@gmail.com)
