+++ test/test_once_storage.rb
@@ -21,7 +21,7 @@
 require 'test/unit'
 require_relative 'test__helper'
 require_relative 'fake_storage'
-require_relative '../objects/once_storage'
+require_relative '../objects/storage/once_storage'
 
 # OnceStorage test.
 # Author:: Yegor Bugayenko (yegor256@gmail.com)
